fixed some bugs (submit order)

+++ app/Http/Controllers/Home/SalesProcess/OrderController.php
@@ -17,33 +17,62 @@ class OrderController extends Controller
 {
     public function submitOrder(SubmitOrderRequest $request)
     {
-        DB::transaction(function () use ($request) {
-            $address = Address::find($request->address);
-            $delivery = Delivery::find($request->delivery_type);
-            $user = Auth::user();
-            $user_obj = ["id" => $user->id, "name" =>  $user->name, "email" =>  $user->email];
-            // check values
-            if ($address->user->id != $user->id) {
-                return back();
-            } elseif ($delivery->status != "active") {
-                return back();
+        $address = Address::find($request->address);
+        $delivery = Delivery::find($request->delivery_type);
+        $user = Auth::user();
+        $user_obj = ["id" => $user->id, "name" =>  $user->name, "email" =>  $user->email];
+        // check values
+        if ($address->user->id != $user->id) {
+            return back();
+        } elseif ($delivery->status != "active") {
+            return back();
+        }
+
+        // calc total price
+        $total_price = 0;
+
+        // find products
+        foreach ($user->cartItems as $cartItem) {
+
+            // check product
+            if ($cartItem->product->status != 'true') {
+                $cartItem->delete();
+                return to_route("home.index")->with("error", "یکی از محصولات سبد خرید شما مشکلی دارد ,لطفا سبد خرید خود را بررسی نمایید");
+            } elseif ($cartItem->color_id != null && $cartItem->product->colors->where("id", $cartItem->color->id)->first() == null) {
+                $cartItem->delete();
+                return to_route("home.index")->with("error", "یکی از محصولات سبد خرید شما مشکلی دارد ,لطفا سبد خرید خود را بررسی نمایید");
+            } elseif ($cartItem->guarantee_id != null && $cartItem->product->guarantees->where("id", $cartItem->guarantee->id)->first() == null) {
+                $cartItem->delete();
+                return to_route("home.index")->with("error", "یکی از محصولات سبد خرید شما مشکلی دارد ,لطفا سبد خرید خود را بررسی نمایید");
+            } elseif ($cartItem->product->marketable != 'true' || $cartItem->product->marketable_number < $cartItem->number) {
+                return to_route("home.salesProcess.myCart")->with("error", "یکی از محصولات سبد خرید شما ناموجود است ,لطفا سبد خرید خود را بررسی نمایید");
             }
 
             // calc total price
-            $total_price = 0;
+            $total_price += $cartItem->totalPrice();
+        }
 
-            foreach ($user->cartItems as $cartItem) {
-                $total_price += $cartItem->totalPrice();
-            }
+        // find tracking_id
+        $user_order = $user->orders()->where("payment_status", "unpaid")->where("status", "not_confirmed")->where("delivery_status", "unpaid")->first();
 
+        if ($user_order != null) {
+            $tracking_id = $user_order->tracking_id;
+        } else {
+            $tracking_id = rand(11111111, 99999999);
+        }
+
+
+        DB::transaction(function () use ($user, $address, $delivery, $tracking_id, $total_price, $user_obj) {
             // find or create Order
             $order = Order::updateOrCreate(
-                ['user_id' => $user->id, "payment_status" => "unpaid", "status" => "not_confirmed"],
+                ['user_id' => $user->id, "payment_status" => "unpaid", "status" => "not_confirmed", "delivery_status" => "unpaid", "tracking_id" => $tracking_id],
                 ['address_id' => $address->id, 'delivery_id' => $delivery->id, "total_price" => $total_price + $delivery->price, "user_obj" => $user_obj, "address_obj" => $address, "delivery_obj" => $delivery]
             );
 
-            // get Products
-            $products = [];
+            // detach old products
+            $order->products()->detach();
+
+            // find product and attach
             foreach ($user->cartItems as $cartItem) {
                 if ($cartItem->color != null) {
                     $color_price = $cartItem->product->colors->where("id", $cartItem->color_id)->first()->pivot->price;
@@ -57,7 +86,7 @@ class OrderController extends Controller
                     $guarantee_price = null;
                 }
 
-                $products[$cartItem->product->id] = [
+                $order->products()->attach($cartItem->product->id, [
                     'color_name' => $cartItem->color->name ?? null,
                     "color_hex_code" => $cartItem->color->hex_code ?? null,
                     "color_price" => $color_price,
@@ -67,10 +96,8 @@ class OrderController extends Controller
                     "number" => $cartItem->number,
                     "total_price" => $cartItem->totalPrice(),
                     "product_obj" => json_encode($cartItem->product->toArray()),
-                ];
+                ]);
             }
-
-            $order->products()->sync($products);
         });
 
         // redirect to the payment page
@@ -80,9 +107,9 @@ class OrderController extends Controller
     // Payment Page
     public function paymentPage()
     {
-        $order = Auth::user()->orders->where("payment_status", "unpaid")->where("status", "not_confirmed")->first();
+        $order = Auth::user()->orders->where("payment_status", "unpaid")->where("status", "not_confirmed")->where("delivery_status", "unpaid")->first();
         if ($order == null) {
-            return to_route("home.index");
+            return to_route("home.salesProcess.myCart")->with("error", "لطفا مجدد تلاش کنید");
         }
 
         $total_price = 0;
@@ -96,43 +123,73 @@ class OrderController extends Controller
     public function payment(PaymentRequest $request)
     {
         $payment_type = $request->payment_type;
-        $order = Auth::user()->orders->where("payment_status", "unpaid")->where("status", "not_confirmed")->first();
+        $order = Auth::user()->orders->where("payment_status", "unpaid")->where("status", "not_confirmed")->where("delivery_status", "unpaid")->first();
+        $cartItems = Auth::user()->cartItems;
+
+        // get totalPrice
+        $totalPrice = 0;
 
         // last check for cartItems
-        foreach (Auth::user()->cartItems as $cartItem) {
+        foreach ($cartItems as $cartItem) {
+            // check product
             if ($cartItem->product->status != 'true') {
                 $cartItem->delete();
-                return back()->with("error", "یکی از محصولات سبد خرید شما مشکلی دارد ,لطفا سبد خرید خود را بررسی نمایید");
+                return to_route("home.index")->with("error", "یکی از محصولات سبد خرید شما مشکلی دارد ,لطفا سبد خرید خود را بررسی نمایید");
             } elseif ($cartItem->color_id != null && $cartItem->product->colors->where("id", $cartItem->color->id)->first() == null) {
                 $cartItem->delete();
-                return back()->with("error", "یکی از محصولات سبد خرید شما مشکلی دارد ,لطفا سبد خرید خود را بررسی نمایید");
+                return to_route("home.index")->with("error", "یکی از محصولات سبد خرید شما مشکلی دارد ,لطفا سبد خرید خود را بررسی نمایید");
             } elseif ($cartItem->guarantee_id != null && $cartItem->product->guarantees->where("id", $cartItem->guarantee->id)->first() == null) {
                 $cartItem->delete();
-                return to_route("home.salesProcess.myCart")->with("error", "یکی از محصولات سبد خرید شما مشکلی دارد ,لطفا سبد خرید خود را بررسی نمایید");
+                return to_route("home.index")->with("error", "یکی از محصولات سبد خرید شما مشکلی دارد ,لطفا سبد خرید خود را بررسی نمایید");
             } elseif ($cartItem->product->marketable != 'true' || $cartItem->product->marketable_number < $cartItem->number) {
                 return to_route("home.salesProcess.myCart")->with("error", "یکی از محصولات سبد خرید شما ناموجود است ,لطفا سبد خرید خود را بررسی نمایید");
             }
+
+            $totalPrice += $cartItem->totalPrice();
         }
 
+        // check delivery price
+        if ($order->delivery->price != $order->delivery_obj["price"]) {
+            return to_route("home.salesProcess.myCart")->with("error", "قیمت روش ارسال مورد نظر تغییر کرده است، لطفا دوباره تلاش کنید");
+        } elseif ($order->delivery->status != "active") {
+            return to_route("home.salesProcess.myCart")->with("error", "روش ارسال مورد نظر غیر فعال شده است، لطفا دوباره تلاش کنید");
+        }
+
+        // check products number
+        if ($cartItems->count() != $order->products->count()) {
+            dd($cartItems->count(), $order);
+            return to_route("home.salesProcess.myCart")->with("error", "لطفا روند ثبت سفارش خود را تکرار کنید");
+        }
+
+        // check final price
+        if (($order->delivery->price + $totalPrice) != $order->total_price) {
+            return to_route("home.salesProcess.myCart")->with("error", "لطفا روند ثبت سفارش خود را تکرار کنید");
+        }
 
         if ($payment_type == 2) {
-            Payment::create([
-                "order_id" => $order->id,
-                "amount" => $order->total_price,
-                "status" => "cash",
-            ]);
+            DB::transaction(function () use ($order, $cartItems) {
+                Payment::create([
+                    "order_id" => $order->id,
+                    "amount" => $order->total_price,
+                    "status" => "cash",
+                    "payment_status" => "cash_payment"
+                ]);
 
-            // update order
-            $order->update([
-                "status" => "confirmed"
-            ]);
+                // update order
+                $order->update([
+                    "delivery_status" => "processing"
+                ]);
 
-            // delete cartItems
-            foreach (Auth::user()->cartItems as $cartItem) {
-                $cartItem->delete();
-            }
+                // delete cartItems and update product
+                foreach ($cartItems as $cartItem) {
+                    $product = $cartItem->product;
+                    $product->increment('sold_number', $cartItem->number);
+                    $product->decrement('marketable_number', $cartItem->number);
+                    $cartItem->delete();
+                }
+            });
 
-            return to_route("home.profile.myProfile.index")->with("success", "سفارش شما با موفقیت ثبت شد");
+            return to_route("home.profile.myOrders.index")->with("success", "سفارش شما با موفقیت ثبت شد");
         }
     }
 }
